feat: implemented error handling; move routes for a new package

+++ models/user.go
@@ -4,7 +4,7 @@ import "github.com/google/uuid"
 
 type User struct {
 	ID          uuid.UUID `json:"id"`
-	Age         int       `json:"age"`
+	Age         int8      `json:"age"`
 	PhoneNumber string    `json:"phone_number"`
 	Email       string    `json:"email"`
 	Address     Address   `json:"address"`
@@ -12,7 +12,7 @@ type User struct {
 
 type Address struct {
 	Street  string `json:"street"`
-	Number  int    `json:"number"`
+	Number  int8   `json:"number"`
 	City    string `json:"city"`
 	State   string `json:"state"`
 	Country string `json:"country"`
